[优化]UiStateTestCustomInt扩大参数范围

+++ annotation/src/main/java/com/mock/annotation/custom/UiStateTestCustomInt.kt
@@ -7,5 +7,5 @@ package com.mock.annotation.custom
 @Retention(value = AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.FIELD)
 annotation class UiStateTestCustomInt(
-    val customInt: Int
+    val customInt: Long
 )
